add cssVar support, move types to separate file, export consts
