+++ packages/schema-extract/test/test.spec.ts
@@ -24,7 +24,7 @@ describe('Stylable JSON Schema Extractor', () => {
         it('schema with a element', () => {
             const res = extractSchema('Comp{}', '/entry.st.css', '/', path);
 
-            expect(res.properties!.Comp).to.flatMatch({
+            expect(res.properties!.Comp).to.eql({
                 $ref: 'stylable/element'
             });
         });
@@ -32,8 +32,16 @@ describe('Stylable JSON Schema Extractor', () => {
         it('schema with a var', () => {
             const res = extractSchema(':vars { myVar: red; }', '/entry.st.css', '/', path);
 
-            expect(res.properties!.myVar).to.flatMatch({
+            expect(res.properties!.myVar).to.eql({
                 $ref: 'stylable/var'
+            });
+        });
+
+        it('schema with a css var', () => {
+            const res = extractSchema('.root { --myVar: red; }', '/entry.st.css', '/', path);
+
+            expect(res.properties!['--myVar']).to.eql({
+                $ref: 'stylable/cssVar'
             });
         });
     });
